Make DockerCompose an interface
